Persist player state and load it server-side, plus bugfixes so we more reliably update our display on login/logout, plus name display in map view.

+++ Shared/UnDecidedPlayer.h
@@ -13,7 +13,7 @@
 
 @property (copy) NSString *userName;
 @property (assign) NSPoint playerPosition;
-@property (assign) uint32_t costumeID;
-@property (assign) uint32_t animationID;
+@property (assign) NSInteger costumeID;
+@property (assign) NSInteger animationID;
 
 @end
